requirements.txt
torchvision
torch
pyquaternion
pyyaml
rospkg
pexpect
mujoco
dm_control
opencv-python
matplotlib
einops
packaging
h5py
ipython
cvxpy
einops
numba
tensorboard
colorama
safetensors
numpy-quaternion
